split smoke and auth integration e2e suites
