+++ e2e/anonymous-smoke.spec.ts
@@ -29,10 +29,10 @@ async function resolveAuthPageState(
   return mainText.includes("Sign in unavailable") ? "unavailable" : "configured";
 }
 
-test.describe("anonymous routes", () => {
+test.describe("anonymous auth surface", () => {
   test.use({ storageState: { cookies: [], origins: [] } });
 
-  test("protected pages redirect to sign in", async ({ page }) => {
+  test("protected pages redirect to the login surface", async ({ page }) => {
     await page.goto("/players");
 
     await expect(page).toHaveURL(/\/login$/);
@@ -46,7 +46,7 @@ test.describe("anonymous routes", () => {
     await expect(page.locator("body")).toContainText(/Supabase is not configured/i);
   });
 
-  test("login page supports mode switching", async ({ page }) => {
+  test("login page shows a valid anonymous auth surface", async ({ page }) => {
     await page.goto("/login");
     const authPage = page.getByRole("main");
 
@@ -60,21 +60,9 @@ test.describe("anonymous routes", () => {
     await expect(authPage.getByRole("button", { name: "Continue with Google" })).toBeVisible();
     await expect(authPage.getByLabel("Email address")).toBeVisible();
     await expect(authPage.getByLabel("Password")).toBeVisible();
-
-    await authPage.getByRole("button", { name: "Create account" }).click();
-    await expect(authPage.getByRole("heading", { name: "Create account" })).toBeVisible();
-    await expect(authPage.getByLabel("Confirm password")).toBeVisible();
-
-    await authPage.getByRole("button", { name: /already have an account\? sign in/i }).click();
-    await expect(authPage.getByRole("heading", { name: "Sign in" })).toBeVisible();
-
-    await authPage.getByRole("button", { name: "Forgot password?" }).click();
-    await expect(authPage.getByRole("heading", { name: "Reset password" })).toBeVisible();
-    await expect(authPage.getByRole("button", { name: "Send reset link" })).toBeVisible();
-    await expect(authPage.getByLabel("Password")).toHaveCount(0);
   });
 
-  test("recovery link opens reset password mode", async ({ page }) => {
+  test("recovery link lands on a valid reset surface", async ({ page }) => {
     await page.goto("/login#type=recovery&access_token=test&refresh_token=test");
     const authPage = page.getByRole("main");
