+++ package.json
@@ -30,7 +30,9 @@
     "auth:local": "node scripts/auth-local.mjs",
     "proof:capture": "node scripts/screenshot-auth.mjs",
     "proof:publish": "node scripts/proof-publish.mjs",
-    "test:e2e": "playwright test",
+    "test:e2e": "playwright test e2e/anonymous-smoke.spec.ts",
+    "test:e2e:auth": "playwright test e2e/auth-integration.spec.ts",
+    "test:e2e:all": "playwright test",
     "design:lint": "npx @google/design.md lint DESIGN.md",
     "design:spec": "npx @google/design.md spec --rules",
     "design:export": "npx @google/design.md export --format tailwind DESIGN.md",
